tweaked internal-only imports, wall-clean

+++ include/Font.hpp
@@ -28,11 +28,11 @@ namespace libsdlgui
         Font(TTFFont& ttfFont, const std::string& name, uint8_t size, Attributes attributes);
 
         Attributes GetAttributes() const { return m_attribs.Get(); }
-        size_t GetCharSize(char c);
-        size_t GetHeight();
-        size_t GetLineSkipHeight();
+        uint32_t GetCharSize(char c);
+        uint32_t GetHeight();
+        uint32_t GetLineSkipHeight();
         std::string GetName() const { return m_name; }
-        size_t GetTextSize(const char* text);
+        uint32_t GetTextSize(const char* text);
         TTF_Font* GetTtf() const { return m_ttf; }
         uint8_t GetSize() const { return m_size; }
     };
